use the cbindgen crate and some decorators to DRY out the rust/python ffi (#6869)
### Problem

Our FFI between pants python and rust is currently split across:
1. `@ffi.def_extern()`-decorated methods in `native.py`,
2. some C code in literal strings in that file,
3. numerous `#[repr(C)]` and `#[no_mangle]` annotations sprinkled around the rust codebase.

Unless you use a regex-based code search tool, this can make it difficult to understand how to change the FFI. There is duplication of the interface definition in the above ways, which can potentially be a source of subtle-to-not-so-subtle errors.

### Solution

- Introduce the [`cbindgen`](https://github.com/eqrion/cbindgen/) crate to generate all the requisite C header declarations for `native.py` from our Rust sources automagically (I was very impressed).
  - There was some hackery required with `re.sub()`, noted in two new methods named `_hackily_*`.
- Create `_ExternSignature` class and `@_extern_decl` annotation to mark python methods with `cffi` equivalents.
- Move FFI initialization to a new `_FFISpecification` class which injects a `self._ffi` object and has a method to generate FFI declarations for methods marked with `@_extern_decl`, as well as register them on the injected `self._ffi` object.

### Result

Changes to Rust objects and methods on the FFI interface no longer require changing multiple places at once, and the type declarations for FFI methods from python are generated from an `@_extern_decl` annotation at the method definition itself. `cbindgen` can and will e.g. add `const` where necessary, so offloading the duty to the `cbindgen` crate (which is also being used in Servo/WebRender, so I'm not too concerned about maintainability) gives us more than just reducing boilerplate, it also improves correctness.