add a --toolchain-variant option to select the compiler for C/C++ (#6800)
### Problem

We have not previously made it possible for the user to select the compiler to use for C/C++ code (although we have unit tests for both compilers). It is currently hardcoded to use LLVM.

### Solution

- Add a `--toolchain-variant` enum option to `NativeBuildSettings`.
- Convert the subsystem dependency in `NativeTask` on `NativeBuildSettings` to be a global subsystem dependency instead of a scoped dependency.
- Add some boilerplate in `NativeTask` to select the toolchain based on the option.
- Add an info-level log to the compile and link tasks stating the executable name which is being used.
- Expand an integration test to cover each toolchain variant.

### Result

Users can now elect to use `--native-build-settings-toolchain-variant=llvm` to use clang/clang++ for the compiler. The default compiler has been changed to be gcc/g++ (previously `llvm` was the only option).