Enforce interpreter constraints for antlr3 python code. (#6924)
The ANTLR3 python runtime is not python3 compatible. Attempting to use it on python3 results in syntax errors in setup.py, that cause the resolve to fail.

In this change we force python_antlr_library to have a python2 compatibility constraint if using antlr3. We also ensure that this propagates to the python_library that is synthesized from it. Finally, we ensure that in our testprojects all libraries that depend on the antlr library themselves have the necessary constraint.

The result is that even if the default interpreter constraint is set to >=3 (e.g., with the --python-setup-interpreter-constraints cmd line flag or the PANTS_PYTHON_SETUP_INTERPRETER_CONSTRAINTS env var), the per-target constraints will override those settings, and a python2 intepreter will be selected.