Fix Travis environment variable overriding interpreter_selection test's config (#6939)
### Problem
In `test_conflict_via_config()`, we try to constraint Python to `< 2.7`. However, this fails to do anything when running on Python 3, because it's overriden by `ci.sh`'s environment variable `export PANTS_PYTHON_SETUP_INTERPRETER_CONSTRAINTS='["CPython>=3.5,<4"]'`.

This is expected behavior, that the environment variable overrides the command line argument. So, the issue is not with the source code, but this test should also not be failing.

### Solution
Set the integration test to be hermetic.

### Other related interpreter selection changes
* Allow Python 3.7 now that it's been fixed with #6363 
* In CI, require for Python 3 mode Python 3.5+, now that we use xenial (#6885)