Set CI's minimum Python 3 version to 3.6 (#6954)
Python 3.6 provides access to the following features:

- f-strings, `f"hello {name}!"`
- underscores in numeric literals, `1_000_000`
- variable type annotations, `x: int = 5`
- async generators
- async comprehensions
- asyncio now stable API + performance and usability enhancements
- stable dict order
- access to dataclasses backport

Because our Pants interpreter constraint merely requires the presence of a Python 3.6+ interpreter in the Pants user's environment—and does not require them to actually change their code—we decided that the benefits outweigh the loss to compatibility.

For users who do not have Python 3.6+ by default (e.g. Ubuntu Trusty and Xenial), they may use tools like Pyenv to get the interpreter, which is what we do in CI. 

Once we officially drop Python 2 support, we can add a runtime check along with a message to help users get Python 3.6+ in their respective environment.

(Note none of this is necessary yet. This simply changes CI and determines our policy.)