Fix tokenize encoding issue engine parser on Python 3 (#6962)
We were always reading file content as bytes, when in Python 2 the tokenize library does expect bytes but in Python 3 it expects unicode. This led to a MappingError whenever trying to execute any goal that involved reading a BUILD file.

We originally improperly fixed this by having the test code pass unicode, which worked because Python 2 would implicitly coerce the test content into unicode, and Python 3 was using the unicode it needed.

We now properly decode the file to unicode conditional on the interpreter used.