Add support for deprecating scoped SubsystemDependencies (#6961)
### Problem

In order for a `v2` `Goal` (in #6880) to replace a goal that was previously implemented as a collection of v1 `Task`s, we need the ability to declare a deprecated dependency on a `Subsystem` (concretely: `CacheSetup`, in that case).

As explained in the patch, this is different from `Optionable.deprecated_scope` in one fundamental way: with `deprecated_scope` options live in two locations simultaneously (one of which is deprecated), while with a deprecated dependency they live in one location (which is deprecated). This split plays out in the unit tests, which verify that `Optionable.deprecated_scope` allows for splitting of a deprecated scope between two new destinations.

### Solution

Add a `removal_version` to `SubsystemDependency` and `ScopeInfo`, and consume it in `Options` in order to render a warning for a deprecated `ScopeInfo`.

The ability to deprecate a scoped dependency is exposed via `Subsystem.scoped(cls, removal_version=.., removal_hint=..)`.

### Result

Callers can use `Subsystem.scoped(cls, removal_version=.., removal_hint=..)` to deprecate a scoped `Subsystem` dependency. #6880 will be unblocked to introduce a deprecated dependency on `CacheSetup` for v2 goals.