Restrict Python 3 to either 3.6 or 3.7
Failing contrib/python/../test_checkstyle.py tests raised the question of if we want to set an upper limit for compatibility, which is required for those tests to function properly.

I realized an upper limit makes sense. Even though it's not yet released, Pants would not support Python 3.8 at the moment, due to several major deprecations become removals. We should only turn on Python 3.8 support when we are ready to do so (in 1 year at release, hopefully).