Fix test_checkstyle.py interpreter constraint  (#6983)
### Problem
Right now, `test_checkstyle.py` relies on the assumption that we only allow `CPython>=2.7,<3`. 

The tests stop behaving as intended when allowing `CPython>=3.6<4`, as we will be doing in https://github.com/pantsbuild/pants/pull/6959. This is because the tests are testing how we handle a whitelist of `CPython>=3.6`, but when we allow this by default, the whitelist no longer matters because the linter will already be working with that interpreter.

### Solution
Constrain the test targets to `CPython>=3.4,<3.6`, as we no longer support these Python 3 versions, yet they are still meaningful versions. When the whitelist is turned on, the tests will run the linter using either Python 3.4 or 3.5.

Note that constraining to `CPython>=3.8` does not work, as Python 3.8 is not yet available so the tests white listing that interpreter will fail due to not finding a corresponding interpreter.

#### Other solution attempted
I first tried to constrain `PythonSetup.global_instance().interpreter_constraints` to only allow Python 2.7, as we do now. 

I got the constraint to work, but the tests were still failing. I found that this is due to the function `PythonSetup.compatibility_or_constraints`, which no matter what uses the passed compatibility constraints when given, rather than first trying to read the target's compatibility. So, the Python 3 targets were improperly resolving as Python 2 compatible.