Replace deprecated cgi.escape() with html.escape() (#6986)
### Problem
`cgi.escape()` has been deprecated since Python 3.2 and it will be removed in Python 3.8. It results in a DeprecationWarning whenever running Pants with Python 3.

### Solution
Allow Pystache to use its default of `html.escape`, which we're still able to access in Python 2 thanks to the `future` backports.

The only behavior difference between what we had vs. `html.escape` is that single quotes are now escaped, which is good for security.