Make UI Per-session instead of per-request (#6827)
### Problem

Currently, the V2 UI works on a per-call-to-`Scheduler::execute` session, in particular every instance of UI is tied to an `ExecutionRequest`. In the V2 model, we have an `ExecutionRequest` per `@console_rule` ran, so if we were to run `./pants --v2 --no-v1 compile test`, two different UIs would be created.

### Solution

The first strep to addressing that is to tie a particular UI to a `Session` instead of an `ExecutionRequest`, since `Sessions` are designed to be "per pants invocation", as opposed to `ExecutionRequest`s.

### Result

`Session`s now contain an instance of the `EngineDisplay`, which can be activated as usual with `./pants --v2 --no-v1 --v2-ui <rules (e.g. list ::)`