A preliminary bootstrapping CI stage, to prevent repeated work in each CI shard. (#7012)
A single job in the bootstrap phase builds a pants.pex and stores it in S3. The jobs in the test stage then retrieve that pants.pex and use it instead of bootstrapping their own. 

This change also splits our CI test stage into more shards, and gives each shard a unique `CACHE_NAME` env var, so that they don't attempt to read/write the same cache (which is inadvisable, according to Travis).  To make this easier and DRYer, we switch from a directly-authored .travis.yml to generating one from a mustache template.  We add a pre-commit check to ensure that the .travis.yml has been correctly generated.

Note that to facilitate pull requests, this change requires a public S3 bucket (`ci-public.pantsbuild.org') with global write permissions. This is obviously a little risky. We mitigate abuse risks thus:

- The bucket has versioning turned on, so that we can detect if anyone overwrites an existing `pants.pex`. 
- The bucket expires all versions of an object after 1 day.

Abuse is still possible, but this seems like a calculated risk worth taking, since the benefits of faster CI are significant.