Add support for consuming Subsystems from @rules (#6993)
### Problem

#6872 laid groundwork for consuming `(Scoped)Options` from `@rules`, but did not actually expose a facility for a `@(console_)rule` to consume a `Subsystem`.

### Solution

Expose `ScopedOptions` parsing to `@console_rules`, and add a helper method that creates an `@rule` to construct a `Subsystem`. Finally, use this facility to have `v2` `list` consume a `Subsystem` that holds options that define (most) of the behaviour that would allow it to replace the `Task`/`v1` implementation of `./pants list`.

### Result

`@rules` and `@console_rules` can consume `Subsystems`. This is a prerequisite for #6880, which will build atop this facility to fully support `v2`-only goals, and replace `v1` `list` with `v2` `list`.