Validate yield statements in rule bodies to remove ambiguity about returning (#7019)
### Problem

See discussion in #7015. `yield` expressions are currently mildly ambiguous in `@rule` bodies :
```python
@rule(A, [])
def f():
  yield Get(B, C, C()) # yield to the engine to get a B, and discard it
  yield A() # yield to the engine and return
```

### Solution

Disallow `yield` statements without an assignment, except when the `yield` means to exit the `@rule` body and return control to the engine. This is done by keeping a set of `yield` expressions that appear in assignment statements while visiting the AST of the `@rule` body, and when a `yield` is found which isn't assigned to anything, ensuring that it is the last statement in a statement body, so that nothing can come after it. This does not affect the current parsing of `Get`s.

### Result

```python
@rule(A, [])
def f():
  yield Get(B, C, C()) # this raises an exception during @rule parse time
  _ = yield Get(B, C, C()) # this is fine
  yield A() # this is unambiguously a call to exit the rule body and return control to the engine
```

This works in if/try/for blocks as well.