[compile.rsc] fix key error; ensure java compiles get necessary zinc scala deps (#7038)
### Problem
- the dependency list for java compiles used the wrong keys in some cases
- collecting the invalid deps for the graph for java zinc compiles requires traversing transitive scala deps

### Solution
- add a helper for metacp deps
- override collecting invalid deps to ensure the walk recurses when the compile target has java deps and the current is scala

### Description

For a plain zinc compile, all of the work has the same flavor, so we can assume that only depending on a direct dependency will give us the outputs needed. But, the Rsc graph has multiple flavors of nodes, so to ensure the right dependency flavors have already been calculated, we need to traverse more of the invalidated graph.