Setup our own configparser backport to only use backport with Py2
The configparser library has a deprecation warning in Python 3. They haven't updated the library in 2 years, so there is little hope this will be addressed. Further, upon reading their documentation, they state you should not use the backport when using Python 3.

So, we instead create our own compat file that will use the backport with Py2 and will use stdlib with Py3.