More accurate wording
Thanks for the clarification John.

Also using the `try except` compatibility idiom now instead of the `if PY3` idiom. If I understand correctly, this library will always need to be executable by a Python 2 process, right? For example we can't use Py3-only syntax in it? If that's the case, then the `try except` idiom is better to minimize the risk of us accidently removing Py2 collection support once we drop Py2 internally.