Improve symlink errors (#7054)
### Problem

When a glob attempted to match something outside the buildroot, the resulting error did not include a useful form of the erroneous path. And additionally, when a symlink was the source of that glob, we would also not render the symlink. See #7051.

### Solution

Render a user-friendly form of the relevant path in the error when we detect a glob the exits the buildroot. Additionally, implement `Display` for `Failure`, and concatenate information about the symlink we were expanding when we fail to expand a symlink.

### Result

The repro from #7051 renders as:
```
./pants --no-print-exception-stacktrace list testprojects/badsymlink::
timestamp: 2019-01-09T20:56:38.674131
Exception caught: (pants.build_graph.address_lookup_error.AddressLookupError) (backtrace omitted)
Exception message: Build graph construction failed: ExecutionError 1 Exception encountered:
  Exception: PathGlobs expansion failed: While expanding link "testprojects/badsymlink/subdir/bad_symlink": Globs may not traverse outside of the buildroot: "testprojects/badsymlink/subdir/../../../../foo"
```

Fixes #7051.