Fix sharing env values & other yml issues
YAML has no way to merge lists. Instead, we exploit Travis's ability to define multiple environment variables in the same line. So, we simply move the variables to one line and assign that to an anchor.

FYI inputting `.travis.yml` into https://yamlvalidator.com is extremely helpful to debug things and see the corresponding JSON.