Base records only add their new info
This is an attempt to get the OSX Py3 shard to work. Instead of records like base_osx_test_config extending base_osx_config, we leave it to the py2 and py3 versions to choose which base setup to extend.

This is an attepmt to fix a suspected merge conflict for the Py3 OSX engine shard. I suspect py3_osx_config is being overriden by base_osx_config. We want py3_osx_config to win out.

More philosophically, because the `base` records are not adequate to stand on their own, they should solely include new information, and not be extending other records, where possible.