Try to set up Docker to build Py3
TIL that Centos6 does not by default have python3, let alone 3.6 or 3.7. Nor can you download it from Yum. So, we must build from source.

This is to fix an issue with the Py3 build engine shard not being able to discover python3 on its path.

Also remove PANTS_PYTHON_SETUP_INTERPRETER_CONSTRAINTS entry because we only want this constrained in `ci.sh`, not in Docker!