Clone pyenv directly on osx engine shard
Allows us to get a modern version of Py3.6 without having to call `brew update`. Note we only do this for the OSX engine shard, and not the default py3_osx_config, because those have a more modern osx_image so can use the default brew-installed pyenv; this allows us to avoid the cost of having to git clone pyenv for those shards.

Thanks John for the suggestion!