Fix setup_py.py unicode issue
We must always write to the outstream with bytes, because `rwbuf` uses bytes and we decided to try to unify our treatment of outstreams (e.g. use sys.stdout in Py2 and equivalent sys.stdout.buffer in Py3).

This issue was causing the build wheels jobs to fail and the contrib tests to fail.

This will be broken out into a separate PR, but is being tested here that it works.