Constrain blacklist shards interpreter to Py3
The underlying PEX is Py2, but subprocesses Py3. This is exactly how we have CI on master setup. Without this commit, we'd be losing some of the Py3 compatibility CI gains we had made before.

The cron job will run purely Py2 pex with Py2 subprocesses to check for regressions.