Revert "Deduplicate py3_osx_config via new mustache file"
Sad, actually we can't deduplicate this :( The reason is that multiple records extend py3_osx_config and py3_osx_test_config. We need some way for osx_test_config to refer to one single entry in the list, as it's impossible to merge lists, so we do need that to all be on one line.