Actually rebuild engine when changing interpreter
Leaving off the `cargo:` part meant this line was ignored. Now we can consistently switch between `./pants` and `./pants3`, and we won't get the _Py_Dealloc issue. Note that it does not always rebuild, as we seem to cache the interpreter-specific native_engine.so after the first time it's built (in ~/.cache/pants/bin/native-engine), which is neat! So we can safely switch between `./pants` and `./pants3`, without having to wait for it to always recompile!

I also checked to make sure it's idempotent, i.e. it doesn't recompile when you keep doing `./pants3` or keep doing `./pants`.

This was tested a few times on macOS and Ubuntu.