Properly render `\n` with exiter in Py3
When ran with Py3, stacktraces would not properly render new lines and would instead print them as literals. This is because the stacktrace was being printed as a byte string (with a b'' prefix).

The issue is from using `print` with bytes. In a Py3 REPL, `print(b"test", file=sys.stderr.buffer)` will raise a TypeError, even though it looks like it should work.

Instead, we must use sys.stdout.buffer.write. The only meaningful difference is that print() adds a new line at the end, so we manually add that ourselves.

Note certain exceptions are still printing the new line literal.. This only fixes one of the sources of the problem.