Refix bad lint
Originally I was going to wait to have this fix be a part of #7067, but we need to include it now to get the linting shard to pass.

I have no idea why the linter passed all this time on master, yet now fails when I didn't even touch source code.. Alas.