Add specific copyright year check on newly added python files (#7066)
### Problem

*See [this slack conversation](https://pantsbuild.slack.com/archives/C046T6T9U/p1547163296121700).*

When people add new python files to pants, they (I) can sometimes add files with the wrong copyright year (the current year), by copy and pasting from another file. This is fine, but can be confusing when looking through git history, and while it can be caught in review, this check happens to be automateable.

### Solution

- Use `git diff --cached` to obtain newly created files added to the index and check their copyright year specifically in `check_header_helper.py`.
- Some other edits to `check_header_helper.py` to support this (raising an exception instead of returning a boolean for a failed header check).
- Leave a TODO linking to #7068 to make more of the pre-commit checks act on files in the index instead of the worktree.
- Allow exporting `IGNORE_ADDED_FILES` to avoid this extra checking entirely.

### Result

Something automateable is automated!