use the asttokens 3rdparty lib to make @rule parsing errors very smooth (#7023)
### Problem

As @stuhood and @baroquebobcat [pointed](https://github.com/pantsbuild/pants/pull/7019#discussion_r245098273) [out](https://github.com/pantsbuild/pants/pull/7019#discussion_r245171933) in #7019, it seems a little silly to only be able to provide the location of the start of the offending `@rule` for `yield` statements which would lead to ambiguity (see #7019), instead of the actual `yield` statement to blame. In that PR, we punted and used `ast.dump()`, but in the meantime I was able to find a 3rdparty library ([`asttokens`](https://github.com/gristlabs/asttokens)) which performs the appropriate bookkeeping to get the correct source text as well as line and column offsets of individual nodes within an [`ast`](https://docs.python.org/2/library/ast.html) parse tree.

If you are interested in why this isn't trivial, I recommend trying an `ast.dump(node, include_attributes=True)` and trying to hook up the numbers yourself (and if it is obvious, let me know, but I don't think this particular 3rdparty lib is too difficult of a dependency to bring in).

### Solution

- Add `asttokens` 3rdparty requirement.
- Use `asttokens.ASTTokens` and `asttokens.LineNumbers` to get the text and position of the offending `yield` statement in the original file.
- Re-add indentation that we had to remove to parse the `@rule` body in the first place so that our error message is exactly what the source text is.

### Result

An invalid rule like:

```python
@rule(A, [])
def f():
  yield Get(SomeProduct, SomeInput, some_input)
  yield A()
```

Will now raise an error like:

```
In function f: yield in @rule without assignment must come at the end of a series of statements.

A yield in an @rule without an assignment is equivalent to a return, and we
currently require that no statements follow such a yield at the same level of nesting.
Use `_ = yield Get(...)` if you wish to yield control to the engine and discard the result.

The invalid statement was:
test_rules.py:45:2
  yield Get(SomeProduct, SomeInput, some_input)

The rule defined by function `f` begins at:
test_rules.py:43:0
@rule(A, [])
def f():
```

I can now jump to a parse error (an invalid `yield` statement, as above) and hit the return key in emacs's compilation-mode to jump again to the precise source text which caused the error. This makes the prospect of adding more `ast` magic in the future (or not, and continuing to use what was added in #7019) in a highly structured and specific way much more plausible and understandable to users not incredibly well-versed in the semantics of the v2 engine.