Fix threading issue with report.py when using Py3
The Py3 threading library has an attribute called _stop() for Events. Our implementation in report.py was thus overriding the defauult, and resulting in the error `TypeError: 'Event' object is not callable` whenever running a test with `./pants3`.

Rename `self._stop` to `self._stopper`.

See https://stackoverflow.com/questions/323972/is-there-any-way-to-kill-a-thread/325528#325528.