Fix threading issue with report.py when using Py3 (#7085)
### Problem
The Py3 threading library has an attribute called `_stop()` for` Event`s. Our implementation in `report.py` was thus overriding the default, and resulting in the error `TypeError: 'Event' object is not callable` whenever running a test with `./pants3`.

### Solution
Rename `self._stop` to `self._stopper`.

See https://stackoverflow.com/questions/323972/is-there-any-way-to-kill-a-thread/325528#325528.