refactor NativeExternalLibraryFetch to be a SimpleCodegenTask  (#7060)
### Problem

We want to allow the use of the existing `remote_sources()` target and `DeferredSourcesMapper` task to orchestrate extracting resources such as C++ headers and libraries from python wheels in #7046. This approach requires creating a target type which the `dest=` kwarg of the `remote_sources()` target populates (`packaged_native_library()` in this PR).

Since we would already be requiring one form of generated targets, it occurred to me that we could use the same approach for `NativeExternalLibraryFetch`, and that `packaged_native_library()` would be a really neat "intermediate representation" for 3rdparty native code which may come from many different sources -- see more exposition in #7046. A user can also, for example, use `packaged_native_library()` for checked-in libraries, which wasn't possible before.

In an abstract sense, since the layout of Conan packages is [not strictly defined (see the "Note" section)](https://docs.conan.io/en/latest/using_packages/conanfile_txt.html#imports) in the way that e.g. jars are, it seems to make sense to wrap it in another target type which further specifies the relative include and lib directories, since those can be set per-package. Having this additional level of indirection allows us to centralize the logic of searching through downloaded packages (a wheel, or a Conan package) for native resources such as headers and libraries.

### Solution

- Allow specifying `include_relpath` and `lib_relpath` in `ConanRequirement`, and deprecate the use of bare strings in the `packages=` kwarg of `external_native_library()`.
- Introduce `packaged_native_library()` as an intermediate representation for 3rdparty headers and libraries, and remove `NativeExternalLibraryFiles`.
- Introduce `NativeTask.external_dependent_constraint` and `NativeTask.packaged_native_deps()` to iterate over the packaged dependencies in native compile and link tasks without any knowledge of the task that produces these packaged dependencies.
- Convert `NativeExternalLibraryFetch` to be a `SimpleCodegenTask` (it *is* pretty simple!) producing `PackagedNativeLibrary`s that the compile and link tasks can now consume.
- Remove an awkward Conan invocation when bootstrapping the pex by using a `remotes.txt` file to populate the remotes in `CONAN_USER_HOME` from the `--conan-remotes` option.
- Introduce the `PythonToolInstance.run_with()` contextmanager to read from the stdout of an invoked python tool while still using the great workunit interface to display stdout/stderr to the user on failure.

### Result

`ConanRequirement`s can now specify the relative paths to the include and lib directories, which are [determined by convention and not enforced](https://docs.conan.io/en/latest/using_packages/conanfile_txt.html#imports). This can be lowered to an instance of the new `PackagedNativeLibrary` target, a low-overhead way to bundle together the many inconsistencies that define the current state of C/C++ 3rdparty packaging. On the consumption end, native compile and link tasks don't depend on a single product from a single task, and the way is paved to introduce native code from python wheels with very low overhead in #7046.