Make wheels shards universal & add osx_image arg
It appears the "build wheels" shards don't even use the PEX, so they should be universal. We can deal with adding to CI a Py3 shard once we convert release.sh to support Py3.

Also, add back the osx_image=xcode8 arg. Maybe this is the source of all the problems..?