Hotfix for #6981 breaking release.sh (#7091)
## Problem
#6981 led to the `Deploy unstable multiplatform pants.pex` shard failing consistently. The error can be reproduced locally via `./build-support/bin/release.sh -p`. The error message, however, is not very descriptive, even with `set -x` and `PANTS_VERBOSE=9`.

## Solution
The likely culprit is removing `osx_image: xcode8` from the `OSX build wheels` shard. We had this hardcoded before #7091 for a reason, and removing it is probably what messed things up. The removal was not intentional, merely an oversight.

Also, this changes the wheel shards to not have py2 vs py3 setups, as they don't use a PEX so the distinction is not meaningful at the moment.

## Result
The result *cannot* be verified until being merged into master. This is because Travis never allows the `deploy` stage to happen from pull requests, as it's a security concern.