Ignore 3rd party deprecation warnings without a valid fix available
Certain libraries, e.g. Future, have not yet fixed their deprecation warnings, even with people submitting PRs to do so. They may not fix it ever if they are unmantained.

We do not want to show these errors as there is nothing Pants or its users can reasonably do to fix the issue.

So, we filter out problematic libraries.

Note we could alternatively fork the library and fix the warnings, but this is much more complex and less ideal.