Reorder CI based on what's most likely to change from a typical PR (#7104)
Ideally CI fails quickly for faster iteration speed. 

So, we reorder CI around the principle of putting what is most likely to be impacted by a typical PR closer to the front. For example, the `lint` shard often fails and every single PR is liable to failing this test, whereas the `JVM tests` shard is only impacted by a small portion of PRs.

This principle leads to the following shard order:
- Lint
- Clippy
- Build wheels
- Unit tests
- Integration tests (Python 3)
- Integration tests (blacklist)
- Rust tests
- Contrib tests
- OSX sanity checks
- OSX platform specific tests
- JVM tests

Note we make an exception for Build wheels. While it's unlikely a PR will break these jobs, several workflows at Twitter and with our release require wheels to be released, so we keep these close to the top to support those workflows.

Compare [before](https://travis-ci.org/pantsbuild/pants/builds/481470034) and [after](https://travis-ci.org/pantsbuild/pants/builds/481546364).