Watchman event files already unicode if Py3
In Py2, we had to add code to decode all files from a watchman's `event["files"]` (see #3951). This was necessary to fix a unicode bug.

However, in Py3, it appears that this value is already unicode, as we get an error saying `str` does not have an attribute `decode()`.

So, we conditionally check if we need to decode or not based off of the interpreter.