Tighten jvm_platform_analysis tests to expect specific error message
In Py2, the tests were failing for an unintended reason. We expected them to fail with the message "Dependencies cannot have a higher java target level than dependees!", but they were failing with the message "Undefined jvm platform 1.8".

This change ensures the tests are failing for the intended reason.

Note this means the tests will now fail for both Py2 and Py3, not just Py3, as they should.